+++ include/RvR/RvR_rand.h
@@ -31,6 +31,6 @@ int32_t  RvR_rand_game_next_range(RvR_rand_game *game, int32_t min, int32_t max)
 
 void RvR_rand_xor_seed(RvR_rand_xor * xor, uint64_t seed);
 uint64_t RvR_rand_xor_next(RvR_rand_xor * xor);
-int32_t RvR_rand_xor_next_range(RvR_rand_xor * xor, int32_t min, int32_t max);
+int64_t RvR_rand_xor_next_range(RvR_rand_xor * xor, int64_t min, int64_t max);
 
 #endif
